+++ src/grocker/builders/build.py
@@ -41,8 +41,8 @@ def build_root_image(docker_client, config):
 
         # FIXME(fbochu): Replace template by env vars
         helpers.render_template(
-            os.path.join(build_dir, 'apt-repositories.sh.j2'),
-            os.path.join(build_dir, 'apt-repositories.sh'),
+            os.path.join(build_dir, 'provision.sh.j2'),
+            os.path.join(build_dir, 'provision.sh'),
             context,
         )